+++ carp_mobile_sensing/lib/data_managers/local/file_data_manager.dart
@@ -227,6 +227,7 @@ class FileDataManager extends AbstractDataManager {
         flush(activeFile, activeSink);
       });
     });
+    await _sink?.close();
     await super.close();
   }
 }